requirements.txt
Django==2
requests
beautifulsoup4
djangorestframework
psycopg2
django-countries
geoip2
